fix: fix getting player list including empty string
Also removes doubly adding clients. There is a inherent circular
dependency of clients and lobbies. Clients can now be created without a
lobby, but its goroutines will exist if it does not have a lobby
pointer.

There should ideally only be one place where the client should be
created, and that is upon websocket connection. As long as we always set
the lobby ptr immediately after initializing the lobby, we should be ok.

This isn't ideal, I'll keep thinking in the future on how to solve this.